[tpcgen-cli]: extract shared parquet writer
