+++ tpcgen-cli/src/lib.rs
@@ -1,4 +1,5 @@
 //! This library contains both the TPCH and TPCDS command line clients.
 mod logging;
+mod parquet;
 pub mod tpcds_cli;
 pub mod tpch_cli;
